Rename fees module to pricing.
